making unwatchOne more secure
Test if a watcher exist for a given prop on obj. The test was previously only applied in the case no watcher function was given but it is also necessary when iterating over object watchers. This commit prevents unwatchOne to throw an error when it's called by the unwatchAll method.